requirements.txt
load_dotenv
docker
tqdm